REF: Move emulator back into tests package
